# netbox-backend/api/openapi/netbox-go-v1.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "NetBox Go Core Workflow API",
    "version": "1.0.0-pre",
    "description": "Pre-publication exact REST contract for the core-workflow-v1 capability profile."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "TokenAuth": []
    },
    {
      "SessionCookie": []
    }
  ],
...
  "paths": {
...
    "/api/dcim/interface-templates/": {
      "get": {
        "operationId": "listInterfaceTemplates",
        "tags": [
          "dcim"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ordering",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "device_type_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mgmt_only",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterfaceTemplateList"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createInterfaceTemplate",
        "tags": [
          "dcim"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterfaceTemplateWrite"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterfaceTemplate"
                }
              }
            }
          }
        }
      }
    },
    "/api/dcim/interface-templates/{id}/": {
      "get": {
        "operationId": "getInterfaceTemplate",
        "tags": [
          "dcim"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterfaceTemplate"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "replaceInterfaceTemplate",
        "tags": [
          "dcim"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterfaceTemplateWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replaced",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterfaceTemplate"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateInterfaceTemplate",
        "tags": [
          "dcim"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterfaceTemplateWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterfaceTemplate"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteInterfaceTemplate",
        "tags": [
          "dcim"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      }
    },
    "/api/dcim/interfaces/": {
      "get": {
        "operationId": "listInterfaces",
        "tags": [
          "dcim"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ordering",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "device_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "device_name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mgmt_only",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterfaceList"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createInterface",
        "tags": [
          "dcim"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterfaceWrite"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interface"
                }
              }
            }
          }
        }
      }
    },
    "/api/dcim/interfaces/{id}/": {
      "get": {
        "operationId": "getInterface",
        "tags": [
          "dcim"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interface"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "replaceInterface",
        "tags": [
          "dcim"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterfaceWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replaced",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interface"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateInterface",
        "tags": [
          "dcim"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterfaceWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interface"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteInterface",
        "tags": [
          "dcim"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "TokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Token <key>"
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      }
    },
    "schemas": {
...
      "Interface": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "device": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ObjectReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "label": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "value",
              "label"
            ],
            "properties": {
              "value": {
                "type": "string"
              },
              "label": {
                "type": "string"
              }
            }
          },
          "enabled": {
            "type": "boolean"
          },
          "mgmt_only": {
            "type": "boolean"
          },
          "mtu": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "speed": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "duplex": {
            "oneOf": [
              {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "value",
                  "label"
                ],
                "properties": {
                  "value": {
                    "type": "string"
                  },
                  "label": {
                    "type": "string"
                  }
                }
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "display": {
            "type": [
              "string",
              "null"
            ]
          },
          "created": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "last_updated": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "count_ipaddresses": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "InterfaceList": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "count",
          "results"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "next": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "previous": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Interface"
            }
          }
        }
      },
      "InterfaceTemplate": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "device_type": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ObjectReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "label": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "value",
              "label"
            ],
            "properties": {
              "value": {
                "type": "string"
              },
              "label": {
                "type": "string"
              }
            }
          },
          "enabled": {
            "type": "boolean"
          },
          "mgmt_only": {
            "type": "boolean"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "display": {
            "type": [
              "string",
              "null"
            ]
          },
          "created": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "last_updated": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "InterfaceTemplateList": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "count",
          "results"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "next": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "previous": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterfaceTemplate"
            }
          }
        }
      },
      "InterfaceTemplateWrite": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "device_type": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "label": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "mgmt_only": {
            "type": "boolean"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "InterfaceWrite": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "device": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "label": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "mgmt_only": {
            "type": "boolean"
          },
          "mtu": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "speed": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "duplex": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ObjectReference": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "url",
          "display"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "display": {
            "type": "string"
          }
        }
      },
...
    }
  }
}